dfs: sync channels on radar during cac
When getting radar event during cac, a new channel
is selected, and the AP starts on it, without setting
the csa_in_progress field.

This results in other interfaces skip this interface
when considering channel sync.

Add another new_chan_chosen field to indicate a new
channel was already chosen for this interface, and
we can sync to it.

Signed-off-by: Eliad Peller <eliad@wizery.com>