dfs: try syncing to running interfaces
if we got radar event during csa, we might end up with
a running/CACing ap (since the second iface won't align
with the first iface, as its channel (which just got
radar event) is not available).

in this case, we want the (first) iface to align with
the running iface.

note that a similar scenario happens on evert channel
switch (as the interfaces are switched serially), so
add an explicit check to make sure the channel is
available before aligning to it.

Signed-off-by: Eliad Peller <eliad@wizery.com>