Mesh: bug fix - make sure when we disconnect we wait for plink_close to be received
When leaving mesh we are not waiting for the other side to respond to our
close command and tear down the interface.
This fix waits for 300ms and only then tears down the interface leaving
enough time for the other side to receive our close action and send us
a close action back.

note - there is still an issue with mesh_join - when calling mesh join
we first tear down the last mesh - in that case we won't wait - and this
bug still occurs - needs to be fixed later on.

Signed-off-by: Chen Loewy <c-loewy@ti.com>