Mesh: Can't join other mesh network after joining mesh network
Due to the fix of how to deauthenticate a mesh interface
the problem was we sent a close action and never waited for a response
now we divided it into sending and waiting for the response and only then
deinitializing the interface.

in the select network scenario - this didn't work properly. we had to shorten
the timer time to 100ms from 300ms, change the code in such a way that it will
deinit only after the timer has finished.

fixes: WILINK8_SW_DEV-496 and WILINK8_SW_DEV-502

Signed-off-by: Chen Loewy <c-loewy@ti.com>